// src-srv/protos/service.ts
export interface GetDocumentResponse {
    /**
     * @generated from protobuf field: elephant.repository.Document document = 1;
     */
    document?: Document;
    /**
     * @generated from protobuf field: int64 version = 2;
     */
    version: bigint;
}
// ...
export interface UpdateRequest {
    /**
     * UUID of the document to update.
     *
     * @generated from protobuf field: string uuid = 1;
     */
    uuid: string;
    /**
     * Document version to create.
     *
     * @generated from protobuf field: elephant.repository.Document document = 2;
     */
    document?: Document;
    /**
     * Meta data to associate with the document version.
     *
     * @generated from protobuf field: map<string, string> meta = 3;
     */
    meta: {
        [key: string]: string;
    };
    /**
     * IfMatch is used for optimistic locks. Set to the version that you require
     * to be the current one for the update to be performed, or -1 to only perform
     * the update if the document doesn't already exist.
     *
     * @generated from protobuf field: int64 if_match = 4;
     */
    ifMatch: bigint;
    /**
     * Status updates to perform.
     *
     * @generated from protobuf field: repeated elephant.repository.StatusUpdate status = 5;
     */
    status: StatusUpdate[];
    /**
     * ACL is an ACL list controlling access to the document.
     *
     * @generated from protobuf field: repeated elephant.repository.ACLEntry acl = 6;
     */
    acl: ACLEntry[];
    /**
     * ImportDirective can be used to preserve timestamps and authorship
     * information from originating systems, but requires the "import_directive"
     * scope for use.
     *
     * @generated from protobuf field: elephant.repository.ImportDirective import_directive = 7;
     */
    importDirective?: ImportDirective;
}
/**
 * ImportDirective can be used to preserve timestamps and authorship information
 * from originating systems.
 *
 * @generated from protobuf message elephant.repository.ImportDirective
 */
export interface ImportDirective {
    /**
     * @generated from protobuf field: string originallyCreated = 1;
     */
    originallyCreated: string;
    /**
     * @generated from protobuf field: string originalCreator = 2;
     */
    originalCreator: string;
}
/**
 * UpdateResponse describes the result of an update request.
 *
 * @generated from protobuf message elephant.repository.UpdateResponse
 */
export interface UpdateResponse {
    /**
     * @generated from protobuf field: int64 version = 1;
     */
    version: bigint;
}
/**
 * @generated from protobuf message elephant.repository.ValidateRequest
 */
export interface ValidateRequest {
    /**
     * @generated from protobuf field: elephant.repository.Document document = 1;
     */
    document?: Document;
}
/**
 * @generated from protobuf message elephant.repository.ValidateResponse
 */
export interface ValidateResponse {
    /**
     * @generated from protobuf field: repeated elephant.repository.ValidationResult errors = 1;
     */
    errors: ValidationResult[];
}
/**
 * @generated from protobuf message elephant.repository.ValidationResult
 */
export interface ValidationResult {
    /**
     * @generated from protobuf field: repeated elephant.repository.EntityRef entity = 1;
     */
    entity: EntityRef[];
    /**
     * @generated from protobuf field: string error = 2;
     */
    error: string;
}
// ...
export interface StatusUpdate {
    /**
     * Name of the status to set.
     *
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * Version to set the status for. Set to -1 to signal that the status no
     * longer is valid for the document. Optional if part of a document update
     * request as it then will default to the version being written.
     *
     * @generated from protobuf field: int64 version = 2;
     */
    version: bigint;
    /**
     * Meta data to include with the status update.
     *
     * @generated from protobuf field: map<string, string> meta = 3;
     */
    meta: {
        [key: string]: string;
    };
    /**
     * IfMatch is used for optimistic locks. Set to the status id that you require
     * to be the current one for the status to be set, or -1 to only perform the
     * update if the status hasn't previously been set.
     *
     * @generated from protobuf field: int64 if_match = 4;
     */
    ifMatch: bigint;
}
// ...
export interface ACLEntry {
    /**
     * URI that identifies the party that's granted access.
     *
     * @generated from protobuf field: string uri = 1;
     */
    uri: string;
    /**
     * Permissions given to the grantee.
     *
     * @generated from protobuf field: repeated string permissions = 2;
     */
    permissions: string[];
}
/**
 * Document is the content
 *
 * @generated from protobuf message elephant.repository.Document
 */
export interface Document {
    /**
     * UUID is a unique ID for the document, this can be a random v4
     * UUID, or a URI-derived v5 UUID.
     *
     * @generated from protobuf field: string uuid = 1;
     */
    uuid: string;
    /**
     * Type is the content type of the document.
     *
     * @generated from protobuf field: string type = 2;
     */
    type: string;
    /**
     * URI identifies the document (in a more human-readable way than
     * the UUID)
     *
     * @generated from protobuf field: string uri = 3;
     */
    uri: string;
    /**
     * URL is the browseable location of the document (if any)
     *
     * @generated from protobuf field: string url = 4;
     */
    url: string;
    /**
     * Title is the title of the document, often used as the headline
     * when the document is displayed.
     *
     * @generated from protobuf field: string title = 5;
     */
    title: string;
    /**
     * Content is the content of the documen, this is essentially what
     * gets rendered on the page when you view a document.
     *
     * @generated from protobuf field: repeated elephant.repository.Block content = 6;
     */
    content: Block[];
    /**
     * Meta is the metadata for a document, this could be stuff like
     * open graph tags and content profile information.
     *
     * @generated from protobuf field: repeated elephant.repository.Block meta = 7;
     */
    meta: Block[];
    /**
     * Links are links to other resources and entities. This could be
     * links to categories and subject for the document, or authors.
     *
     * @generated from protobuf field: repeated elephant.repository.Block links = 8;
     */
    links: Block[];
    /**
     * Language is the language used in the document as an IETF language
     * tag. F.ex. "en", "en-UK", "es", or "sv-SE".
     *
     * @generated from protobuf field: string language = 9;
     */
    language: string;
}
/**
 * Block is the building block for data embedded in documents. It is
 * used for both content, links and metadata. Blocks have can be
 * nested, but that's nothing to strive for, keep it simple.
 *
 * @generated from protobuf message elephant.repository.Block
 */
export interface Block {
    /**
     * ID is the block ID
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * UUID is used to reference another Document in a block.
     *
     * @generated from protobuf field: string uuid = 2;
     */
    uuid: string;
    /**
     * URI is used to reference another entity in a document.
     *
     * @generated from protobuf field: string uri = 3;
     */
    uri: string;
    /**
     * URL is a browseable URL for the the block.
     *
     * @generated from protobuf field: string url = 4;
     */
    url: string;
    /**
     * Type is the type of the block
     *
     * @generated from protobuf field: string type = 5;
     */
    type: string;
    /**
     * Title is the title/headline of the block, typically used in the
     * presentation of the block.
     *
     * @generated from protobuf field: string title = 6;
     */
    title: string;
    /**
     * Data contains block data
     *
     * @generated from protobuf field: map<string, string> data = 7;
     */
    data: {
        [key: string]: string;
    };
    /**
     * Relationship describes the relationship to the document/parent
     * entity
     *
     * @generated from protobuf field: string rel = 8;
     */
    rel: string;
    /**
     * Role is used for nuancing and qualifying relationships.
     *
     * @generated from protobuf field: string role = 9;
     */
    role: string;
    /**
     * Name is a name for the block. An alternative to "rel" when
     * relationship is a term that doesn't fit.
     *
     * @generated from protobuf field: string name = 10;
     */
    name: string;
    /**
     * Value is a value for the block. Useful when we want to store a
     * primitive value.
     *
     * @generated from protobuf field: string value = 11;
     */
    value: string;
    /**
     * ContentType is used to describe the content type of the
     * block/linked entity if it differs from the type of the block.
     *
     * @generated from protobuf field: string contentType = 12;
     */
    contentType: string;
    /**
     * Links are used to link to other resources and documents.
     *
     * @generated from protobuf field: repeated elephant.repository.Block links = 13;
     */
    links: Block[];
    /**
     * Content is used to embed content blocks.
     *
     * @generated from protobuf field: repeated elephant.repository.Block content = 14;
     */
    content: Block[];
    /**
     * Meta is used to embed metadata
     *
     * @generated from protobuf field: repeated elephant.repository.Block meta = 15;
     */
    meta: Block[];
}
// ...
class GetDocumentResponse$Type extends MessageType<GetDocumentResponse> {
    constructor() {
        super("elephant.repository.GetDocumentResponse", [
            { no: 1, name: "document", kind: "message", T: () => Document },
            { no: 2, name: "version", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ }
        ]);
    }
    create(value?: PartialMessage<GetDocumentResponse>): GetDocumentResponse {
        const message = { version: 0n };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<GetDocumentResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetDocumentResponse): GetDocumentResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* elephant.repository.Document document */ 1:
                    message.document = Document.internalBinaryRead(reader, reader.uint32(), options, message.document);
                    break;
                case /* int64 version */ 2:
                    message.version = reader.int64().toBigInt();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetDocumentResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* elephant.repository.Document document = 1; */
        if (message.document)
            Document.internalBinaryWrite(message.document, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* int64 version = 2; */
        if (message.version !== 0n)
            writer.tag(2, WireType.Varint).int64(message.version);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message elephant.repository.GetDocumentResponse
 */
export const GetDocumentResponse = new GetDocumentResponse$Type();
// ...
export const UpdateRequest = new UpdateRequest$Type();
// ...
export const ImportDirective = new ImportDirective$Type();
// ...
export const UpdateResponse = new UpdateResponse$Type();
// ...
export const ValidateRequest = new ValidateRequest$Type();
// ...
export const ValidateResponse = new ValidateResponse$Type();
// ...
export const ValidationResult = new ValidationResult$Type();
// ...
export const EntityRef = new EntityRef$Type();
// ...
export const StatusUpdate = new StatusUpdate$Type();
// ...
export const ACLEntry = new ACLEntry$Type();
// ...
export const Document = new Document$Type();
// ...
export const Block = new Block$Type();